+++ src/DataSources/Models/DataSource.cs
@@ -4,7 +4,7 @@ public abstract class DataSource
 {
     public required string CollectionId { get; set; }
     public required string Id { get; set; }
-    public required bool Recursive { get; set; }
+    public bool Recursive { get; set; } = true;
     public required string Path { get; set; }
     public string? FileIgnorePatterns { get; set; }
     public int? IgnoreFileIfMoreThanThisNumberOfLines { get; set; }