+++ src/IngestionOptions.cs
@@ -1,0 +1,8 @@
+﻿using SimpleRag.DataSources.CSharp.Models;
+
+namespace SimpleRag;
+
+public class IngestionOptions
+{
+    public Func<CSharpChunk, string>? CSharpContentFormatBuilder { get; set; }
+}