+++ src/Interfaces/ISingletonService.cs
@@ -1,3 +1,0 @@
-namespace SimpleRag.Interfaces;
-
-public interface ISingletonService;