+++ src/VectorStorage/VectorStoreQuery.cs
@@ -1,11 +1,10 @@
 ﻿using System.Linq.Expressions;
 using Microsoft.Extensions.VectorData;
-using SimpleRag.Interfaces;
 using SimpleRag.VectorStorage.Models;
 
 namespace SimpleRag.VectorStorage;
 
-public class VectorStoreQuery(VectorStore vectorStore, VectorStoreConfiguration vectorStoreConfiguration) : IScopedService
+public class VectorStoreQuery(VectorStore vectorStore, VectorStoreConfiguration vectorStoreConfiguration)
 {
     private bool _creationEnsured;
 
@@ -23,7 +22,7 @@ public class VectorStoreQuery(VectorStore vectorStore, VectorStoreConfiguration 
     }
 
 
-    public async Task<SearchResult> SearchAsync(string searchQuery, int numberOfRecordsBack, Expression<Func<VectorEntity, bool>>? filter) //todo - give back a more complex object
+    public async Task<SearchResult> SearchAsync(string searchQuery, int numberOfRecordsBack, Expression<Func<VectorEntity, bool>>? filter)
     {
         VectorStoreCollection<string, VectorEntity> collection = await GetCollectionAndEnsureItExist();
         await collection.EnsureCollectionExistsAsync();