+++ CHANGELOG.md
@@ -1,17 +1,21 @@
 # Changelog: SimpleRag
 
-## Unreleased
+## Version 0.0.1-preview3 (29th of June 2025)
 - Fixed NuGet Link in ReadMe
+- Removed Interfaces folder (not used)
+- Moved Notification system into Models folder
+- Added TopLevel Ingestion and Search classes for easier usage
+- Switched to slnx file
 
 ---
 
-## Version 0.0.1-preview2
+## Version 0.0.1-preview2 (28th of June 2025)
 - Fixed Changelog path in Nuget Package
 - Added initial data in ReadMe (more will come later)
 
 ---
 
-## Version 0.0.1-preview1
+## Version 0.0.1-preview1 (28th of June 2025)
 - Initial Preview Release (WARINING: Everything is subject to Change untill 1.0.0)
 
 ---